error handling struct improved
